+++ src/components/Buttons.js
@@ -91,6 +91,14 @@ function SubmitButton({ onPress, title }) {
   );
 };
 
+function ReviewButton({ onPress }) {
+  return (
+    <TouchableOpacity style={ReviewButtonstyles.reviewButton} onPress={onPress}>
+      <Text style={ReviewButtonstyles.reviewButtonText}>후기 보기</Text>
+    </TouchableOpacity>
+  );
+};
+
 //active, inactive backgroundColor만 달라서 하나로 통일할지 고민
 const styles = StyleSheet.create({
   activeBottomButton: {
@@ -200,6 +208,22 @@ const styles = StyleSheet.create({
   },
 });
 
+const ReviewButtonstyles = StyleSheet.create({
+  reviewButton: {
+    position: 'absolute',
+    top: 25, // 필요에 따라 위치 조정 가능
+    right: 20,
+    backgroundColor: colors.themeColor,
+    paddingVertical: 8,
+    paddingHorizontal: 25,
+    borderRadius: 20,
+    zIndex: 10, // 항상 위에 표시되도록 설정
+  },
+  reviewButtonText: {
+    ...fontStyles.whiteSemiBold14,
+  },
+});
+
 export {
   BottomButton,
   CreatePostButton,
@@ -209,4 +233,5 @@ export {
   FreeButton,
   NavigateButtonTheme,
   SubmitButton,
+  ReviewButton,
 };
